Fix the error when set vimrc.

+++ install.sh
@@ -10,16 +10,33 @@ git clone https://github.com/vashdawn/dots.git ~/.dotfiles
 # Set /etc/vimrc ------------------------------------
 VIM_CONF=~/.dotfiles/vim/vimrc
 SPACEVIM=~/.dotfiles/vim/.spacevim
-[ -e ~/.vimrc -a ! -e ~/.spacevim ] && bak_old ~/.vimrc && ln -s $VIM_CONF ~/.vimrc
-[ -e ~/.spacevim ] && bak_old ~/.spacevim && ln -s $SPACEVIM ~/.spacevim
+if [ -e ~/.spacevim ]; then
+  bak_old ~/.spacevim && ln -s $SPACEVIM ~/.spacevim
+else
+  [ -e ~/.vimrc ] && bak_old ~/.vimrc
+  ln -s $VIM_CONF ~/.vimrc
+fi
 # [ -e ~/.vimrc  ] && [ -e ~/.spacevim ] || bak_old ~/.vimrc
 
 
 # Set /etc/bash.bashrc --------------------------
 BASH_CONF=~/.dotfiles/bash/bashrc
-[ -e ~/.bashrc ] && bak_old ~/.bashrc
+echo -n 'Do you want to replace the default ".bashrc"?(y/n)'
+while true; do
+  read answer
+  if [ $answer = 'y' -o $answer = 'n' ]
+  then
+    break
+  else
+    echo -n 'Please say yes or no!(y/n)'
+  fi
+done
 
-ln -s $BASH_CONF ~/.bashrc
+if [ $answer = 'y' ]; then
+  [ -e ~/.bashrc ] && bak_old ~/.bashrc
+  ln -s $BASH_CONF ~/.bashrc
+fi
+
 
 
 # Set /etc/zsh/zshrc ------------------------------
@@ -49,24 +66,24 @@ ln -s $IP_KEYBINDINGS $TARGET_IP_KEYBINDINGS
 echo 'Please input your proxy server(x.x.x.x:port): '
 read proxy_server
 if [ -n $proxy_server ]; then
-	wget_proxy="\n
-	https_proxy = http://${proxy_server}\n
-	http_proxy = http://${proxy_server}\n
-	ftp_proxy = http://${proxy_server}\n
-	use_proxy = on\n
-	continue = on\n
-	check_certificate = off\n
-	"
+  wget_proxy="\n
+  https_proxy = http://${proxy_server}\n
+  http_proxy = http://${proxy_server}\n
+  ftp_proxy = http://${proxy_server}\n
+  use_proxy = on\n
+  continue = on\n
+  check_certificate = off\n
+  "
 
-	curl_proxy="\n
-	-L\n
-	proxy = ${proxy_server}\n
-	"
+  curl_proxy="\n
+  -L\n
+  proxy = ${proxy_server}\n
+  "
 
-    echo -e $curl_proxy >> ~/.curlrc  # for curl
+  echo -e $curl_proxy >> ~/.curlrc  # for curl
 
-    echo -e $wget_proxy >> ~/.wgetrc  # for wget
+  echo -e $wget_proxy >> ~/.wgetrc  # for wget
 else
-    echo 'No proxy provided.'
+  echo 'No proxy provided.'
 fi
